split code for clarity reasons
